+++ src/Metrics/MANGFMSEImageToImageMetric/itkMANGFMSE.hxx
@@ -182,7 +182,7 @@ MANGFMSE<TFixedImage, TMovingImage>
 
 
 	derivative=a;
-	for(int p=0;p<derivative.GetSize();p++)
+	for(long unsigned int p=0;p<derivative.GetSize();p++)
 	{derivative[p]= a[p]+ this->m_lambdaDerivative*b[p] + this->m_muDerivative*b[p];}
